Introduce methods.lua, to let one define classes in a Factorio-friendly way
Normally this is easier, but if we want any hope of being able to work with global, we have
to do extra special magic.  Internally this is complex, and also don't overly trust iot to work with global until
it's proven  itself, but externally one can use it easily to get "classes".

The next commit contains an example of how to use it, including declaring for LuaLS