Add support for reading the consumption stats, and clean up the code while we're here
We introduce fa_utils.spacecat, which takes any number of stringifiable things
and converts them into a space-separated list.  This is more efficient and
easier to read than .. (if Lua knows ahead of time it may avoid intermediate
string building), and can get rid of the hard to follow result variables. Then
we use this to easily tweak wording in selection_item_production_stats_info, to
be more brief and to add the consumption (output in mod language) stats.

Note a couple things.  First the output code and input code are the same by one
boolean, thus a quick local closure.  After that and the wording tweak it then
becomes evident that we need internal_name only in two spots, so we get to kill
that variable.  In context we have enough to not use super long variable names,
so we get to also simplify some of those.