When reading tiles, skip robots and remnants ...
When reading any tile, while the entity found is a flying robot or a remnant, and while there is at least one other valid entity, do skip to the next entity. Therefore a robot or a remnant is not read unless it is the only entity on the tile.

References #139. UNTESTED.