Added travel menu search + fixed lingering graphics bug (#233)
Travel menu search works like the other menu search features.

The graphics bug was that after loading a game with a menu left open, unreferenced renders would be stuck in the world and on the player GUI. The solution was to make the renders expire and the graphics be periodically cleaned. This makes it far less needed to press "CONTROL + ALT + R".