Trains: fix reaching stops which aren't at the ends of rails, with tradeoffs
What is going on here is that we broke train stuff in a way that kind of works.
If someone puts their train stations always at the end of tracks, it works.
Otherwise, the station is mysteriously unreachable.  It turns out that this is
because prior to this commit, we accidentally didn't add a condition to the
temporary stop.  As a result the train would touch the station but not break,
then us setting it to manual mode would keep the velocity but not the pathing,
as if the player just held forward.

Unfortunately the problem is that it turns out that our announcement of temporary travel completing apparently depends on this behavior, as does setting the train to manual mode on arrival.  We can't do anything about these in a reasonable amount of time given that we are also tackling 2.0, but more clever ways are possible.  It also, for some reason, will no path if the user travels to the next station in the schedule.

This is "fine" though.  Unlike the current problem of mysteriously unreachable
train stops the workaround is simple and explainable and we can just tell people
we'll work on it after 2.0.  We lose the manual mode thing but trains no longer
go anywhere as long as the player is still in it.  That is in effect a rollback,
from the user's perspective, but in exchange for "you did 3 hours of stuff and
it broke sorry" not happening.