Abstract out sneaking data from the prototype to runtime stage, and use it to fix new scanner when playing with mods that add many resource types
It turns out that there's a length limit on localised_description, but we can shove string pairs in
as long as we never try to print it. Instead of using one prototype, use many, of the form
fa-data-map-mapname-index, where index ranges from 1 to the number of map keys. Also
validate that empty maps were declared.