fa-info, control: take the first step toward localisable ent_info
This is not localisable, but we can reshuffle things: ent_info now returns
a string wrapped in a table, preventing the caller from concatenating, and
read_tile in control.lua now has a read_tile_inner wehich builds a table that the parent read_tile function
concatenates with fa-utils.  As a result
it's fine to start returning tables, and we can begin working top-down to get ent_info up to par.