fa-utils: rename offset_position to offset_position_legacy, and explain that this is because it cannot deal with the 16-direction world
Essentially it's actually "move by tile", not offset by unit vector, and will do
bad things if passed other directions (nor can it be rectified).  We can't kill it because lots of code
wants to move at angles by tile, but we'll have to do something better soon, and continuing to use this is probably not a great plan.

We can rename it back or rename it to something better or... if we decide we do need it.